# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CareerPilot \u2014 Multi-agent AI job discovery pipeline",
    "description": "Three agents (Scout finds company job boards, Analyst scores postings against my resume, Coach identifies recurring skill gaps), LangGraph orchestration, RAG over the job archive, and a hand-labeled evaluation set.\n\nThis is the HTTP layer over that pipeline. It wraps adapters/, filters.py, agents/, orchestrator.py, db.py, extraction.py and rag.py - it does not reimplement any of them.\n\n**The system finds and ranks; the human applies.** Nothing here submits an application to an employer, and no endpoint exists that could - that is a design decision, not a missing feature.\n\nRanking quality is measured against the hand-labeled set rather than assumed. The current result is marginal - stage-1 MRR sits only slightly above the random baseline on a small overlap. See GET /api/meta/evaluation for the numbers and their caveats.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/run": {
      "post": {
        "tags": [
          "run"
        ],
        "summary": "Start Run",
        "description": "Starts the orchestrator in a background thread.\n\nOne at a time: a second call while a run is active returns the running\nrun's id with already_running=True rather than starting a concurrent\none. Two orchestrators would double-spend the same Gemini free-tier\nquota and interleave writes to the same SQLite database.\n\nAn optional body.thread_id forces a fresh graph thread - see\nRunStartRequest for why that escape hatch has to exist.",
        "operationId": "start_run_api_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RunStartRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/resume": {
      "get": {
        "tags": [
          "resume"
        ],
        "summary": "Get Resume",
        "operationId": "get_resume_api_resume_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "resume"
        ],
        "summary": "Preview Resume",
        "description": "Accepts a PDF upload or pasted text. Saves nothing.",
        "operationId": "preview_resume_api_resume_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_preview_resume_api_resume_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumePreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resume/confirm": {
      "post": {
        "tags": [
          "resume"
        ],
        "summary": "Confirm Resume",
        "operationId": "confirm_resume_api_resume_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeConfirmResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/meta/runs": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Run History",
        "description": "Recorded runs, most recent first, each with its per-stage metrics.\n\nOnly runs started through this API appear here - a CLI run spends the\nsame quota but writes no metrics row (see /runtime's caveats), so this\nis a history of API-triggered runs, not of every run that happened.",
        "operationId": "run_history_api_meta_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentMetric": {
        "properties": {
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "companies_checked": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Companies Checked"
          },
          "jobs_retrieved": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jobs Retrieved"
          },
          "jobs_processed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jobs Processed"
          },
          "llm_calls": {
            "type": "integer",
            "title": "Llm Calls",
            "default": 0
          },
          "cache_hits": {
            "type": "integer",
            "title": "Cache Hits",
            "default": 0
          },
          "cache_misses": {
            "type": "integer",
            "title": "Cache Misses",
            "default": 0
          },
          "cache_hit_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Hit Rate"
          },
          "retries": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retries"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          }
        },
        "type": "object",
        "required": [
          "stage"
        ],
        "title": "AgentMetric",
        "description": "Per-stage metrics from one run.\n\n`agent` is null for fetch/filter - they're pipeline stages, not agents.\nOnly the Analyst runs in the nightly path; Scout is on-demand (recorded\nseparately, see /api/companies/scout) and Coach is weekly. Labelling a\nfetch stage with an agent name would misrepresent what ran.\n\n`retries` is null, not 0, when unmeasured. LangGraph's RetryPolicy\nretries a node silently - nothing observable fires - so reporting 0\nwould be a confident claim that no retry happened when the truth is\nthat nobody counted. Null means \"not measured\"."
      },
...
      "Body_preview_resume_api_resume_post": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "type": "object",
        "title": "Body_preview_resume_api_resume_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResumeConfirmRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ResumeConfirmRequest"
      },
      "ResumeConfirmResponse": {
        "properties": {
          "saved": {
            "type": "boolean",
            "title": "Saved"
          },
          "length": {
            "type": "integer",
            "title": "Length"
          },
          "invalidated_cached_analyses": {
            "type": "integer",
            "title": "Invalidated Cached Analyses"
          }
        },
        "type": "object",
        "required": [
          "saved",
          "length",
          "invalidated_cached_analyses"
        ],
        "title": "ResumeConfirmResponse"
      },
      "ResumeExtraction": {
        "properties": {
          "skills": {
            "type": "string",
            "title": "Skills"
          },
          "projects": {
            "type": "string",
            "title": "Projects"
          },
          "extracted": {
            "type": "boolean",
            "title": "Extracted"
          }
        },
        "type": "object",
        "required": [
          "skills",
          "projects",
          "extracted"
        ],
        "title": "ResumeExtraction",
        "description": "What extraction.extract_resume_sections pulled, and whether it\nactually recognized headers or fell back.\n\n`extracted=False` means no \"Technical Skills\"/\"Projects\" header matched\nand the FULL text would be sent to the Analyst instead of a trimmed\nsection. That's surfaced as its own boolean rather than left implicit,\non the same visible-fallback principle the JD extractor follows: a\nsilent fallback looks exactly like a successful extraction from the\noutside, and would quietly change every score."
      },
      "ResumePreview": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "length": {
            "type": "integer",
            "title": "Length"
          },
          "extraction": {
            "$ref": "#/components/schemas/ResumeExtraction"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "invalidates_cached_analyses": {
            "type": "integer",
            "title": "Invalidates Cached Analyses"
          },
          "previous_length": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Length"
          }
        },
        "type": "object",
        "required": [
          "text",
          "length",
          "extraction",
          "invalidates_cached_analyses"
        ],
        "title": "ResumePreview",
        "description": "A candidate resume, not yet saved. `warnings` are the mangled-PDF\ntripwires (unusually long words, low space ratio, big length drop) -\nempty means nothing obvious looked wrong, never that extraction is\nproven clean.\n\n`invalidates_cached_analyses` is how many currently-analyzed jobs would\nneed re-scoring if this were saved: the resume text is part of every\nAnalyst cache key, so changing it doesn't corrupt anything, it just\nstops matching - and the next run spends real quota re-analyzing."
      },
      "ResumeResponse": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "length": {
            "type": "integer",
            "title": "Length"
          },
          "extraction": {
            "$ref": "#/components/schemas/ResumeExtraction"
          },
          "exists": {
            "type": "boolean",
            "title": "Exists"
          }
        },
        "type": "object",
        "required": [
          "text",
          "length",
          "extraction",
          "exists"
        ],
        "title": "ResumeResponse"
      },
...
      "RunHistoryEntry": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "trigger": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "total_llm_calls": {
            "type": "integer",
            "title": "Total Llm Calls",
            "default": 0
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/AgentMetric"
            },
            "type": "array",
            "title": "Stages"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status",
          "started_at"
        ],
        "title": "RunHistoryEntry",
        "description": "One recorded run, with its per-stage detail.\n\nSame AgentMetric shape as AgentsResponse.stages - the history view and\nthe latest-run view render identical rows, so they must not diverge\ninto two subtly different metric types."
      },
      "RunHistoryResponse": {
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunHistoryEntry"
            },
            "type": "array",
            "title": "Runs"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "RunHistoryResponse"
      },
      "RunStartRequest": {
        "properties": {
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          }
        },
        "type": "object",
        "title": "RunStartRequest",
        "description": "Optional body for POST /api/run.\n\nthread_id overrides orchestrator.run_nightly's automatic date-based\nthread selection. It exists because that automatic selection is\ndeliberately idempotent: once `nightly-YYYY-MM-DD` has reached END,\nre-invoking it resumes a completed graph, which LangGraph correctly\nanswers by returning the checkpointed final state without executing a\nsingle node. That is the right behaviour for a 2am cron that might\nfire twice - but from a UI it looks like \"Start run\" did nothing, with\nan empty timeline and no explanation.\n\nPassing an unused thread_id forces a genuinely fresh run. The UI sends\none only when the user explicitly asks to re-run a day that already\ncompleted, so the ordinary path keeps the crash-resume behaviour the\ndate-based ids give it."
      },
      "RunStartResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "already_running": {
            "type": "boolean",
            "title": "Already Running",
            "default": false
          },
          "no_op": {
            "type": "boolean",
            "title": "No Op",
            "default": false
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status"
        ],
        "title": "RunStartResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}